Refactor: Position magnifier with Positioned widget
The magnifier widget, previously padded, is now positioned absolutely
using the `Positioned` widget. This change allows for more precise
placement of the magnifier when a tap event occurs on the chart.

The `margin` property was removed from the `Container` widget, and
the `focalPointOffset` of the `RawMagnifier` was adjusted to
accommodate the new positioning.